consolidate calendar components to separate tsx file
